1) Create separate string parsing routines, mostly parameter array to string and vice versa. I find that this nearly always eliminates hidden bugs caused by incorrect handling of spaces, backslashes, utf-8, etc. I often use spaces in file names and path names and never had problems.
2) Move all system calls (e.g. Popen) to a separate routine. While this
adds code, and is likely not necessary for 10.5 (where subprocess.Popen
is available), I am convinced it is bug-free for 10.4 (while the
emulation of Popen in the code likely won't)

3) make it more object oriented. For example, I created a TexMate class,
which acts as a program. Basically, this chopped the currently very long
main() routines in easy to graps pieces.

4) Remove the direct mapping between texCommand (the argument to
texMate.py) and the action that is taken. I distinguish between
do_latex(), which is called once with "texMate.py latex", and
run_latex() which is called once each time the latex program is run. So
do_latex() may call run_mklatex(); run_latex() once; or run_latex(),
run_bibtex(), run_makeindex(), run_latex(), and run_latex() in succession.

5) Add support to detect which file is currently parsed (see parseLine()
in texparser.py), fixing the links in th Typeset window.

6) support for subdirectories. I even took it so far that the line
%!TEX TS-options = -output-directory "output"
was recognized in the .tex files, and acted upon accordingly. There even
was an error in case bibtex was used in conjunction with \include and
subdirectories. This requires implementation of #1 (string parsing
subroutines).